Convert PhaseParser into hook
