+++ src/interface/report/Results/PhaseSelector.tsx
@@ -1,5 +1,5 @@
 import { Phase } from 'game/raids';
-import { SELECTION_ALL_PHASES, SELECTION_CUSTOM_PHASE } from 'interface/report/PhaseParser';
+import { SELECTION_ALL_PHASES, SELECTION_CUSTOM_PHASE } from 'interface/usePhases';
 import Fight from 'parser/core/Fight';
 import * as React from 'react';
