+++ src/interface/report/index.tsx
@@ -3,6 +3,7 @@ import { wclGameVersionToExpansion } from 'game/VERSIONS';
 import ErrorBoundary from 'interface/ErrorBoundary';
 import makeAnalyzerUrl from 'interface/makeAnalyzerUrl';
 import NavigationBar from 'interface/NavigationBar';
+import { SELECTION_ALL_PHASES } from 'interface/usePhases';
 import Config from 'parser/Config';
 import CharacterProfile from 'parser/core/CharacterProfile';
 import CombatLogParser from 'parser/core/CombatLogParser';
@@ -23,7 +24,7 @@ import EventsLoader from './EventsLoader';
 import FightSelection from './FightSelection';
 import ParserLoader from './ParserLoader';
 import PatchChecker from './PatchChecker';
-import PhaseParser, { SELECTION_ALL_PHASES } from './PhaseParser';
+import PhaseParser from './PhaseParser';
 import PlayerLoader from './PlayerLoader';
 import ReportLoader from './ReportLoader';
 import Results from './Results';
